Conflicts:
	etc/Text/djvAV.fr.text
	etc/Text/djvUI.cs.text
	etc/Text/djvUI.da.text
	etc/Text/djvUI.de.text
	etc/Text/djvUI.el.text
	etc/Text/djvUI.en.text
	etc/Text/djvUI.es.text
	etc/Text/djvUI.fr.text
	etc/Text/djvUI.is.text
	etc/Text/djvUI.it.text
	etc/Text/djvUI.ja.text
	etc/Text/djvUI.ko.text
	etc/Text/djvUI.pl.text
	etc/Text/djvUI.pt.text
	etc/Text/djvUI.ru.text
	etc/Text/djvUI.sv.text
	etc/Text/djvUI.zh.text

+++ lib/djvUI/IconSystem.h
@@ -61,9 +61,9 @@ namespace djv
             virtual ~IconSystem();
 
             static std::shared_ptr<IconSystem> create(const std::shared_ptr<Core::Context>&);
-                
+            
             //! Get an icon.
-            std::future<std::shared_ptr<AV::Image::Image> > getIcon(const std::string &, int size);
+            std::future<std::shared_ptr<AV::Image::Image> > getIcon(const std::string &, float size);
 
             //! Get the cache percentage used.
             float getCachePercentage() const;
